Convert coordinates (strings) to floats
The search API returns a string with coordinates that are passed along to
the OpenLayers animate() method. If these coordinates contain decimals then
animate() would point the user to the location on the map. However, if the
coordinates didn't contain decimals the map would freeze up and eventually
have the browser run out of memory.

Converting the coordinates to an array of floats resolves this issue.